+++ Source/Core/ExtentOutlineGeometry.js
@@ -342,7 +342,7 @@ define([
         var granularityX = (extent.east - extent.west) / (width - 1);
         var granularityY = (extent.north - extent.south) / (height - 1);
 
-        var radiiSquared = ellipsoid.getRadiiSquared();
+        var radiiSquared = ellipsoid.radiiSquared;
 
         extent.getNorthwest(nwCartographic);
         extent.getCenter(centerCartographic);