+++ Source/Scene/CameraController.js
@@ -631,6 +631,11 @@ define([
             newLeft = -maxRight;
         }
 
+        if (newRight <= newLeft) {
+            newRight = 1.0;
+            newLeft = -1.0;
+        }
+
         var ratio = frustum.top / frustum.right;
         frustum.right = newRight;
         frustum.left = newLeft;
@@ -819,9 +824,8 @@ define([
 
     defineProperties(CameraController.prototype, {
         /**
-         * The camera heading in radians.
-         * @memberof CameraController
-         *
+         * Gets or sets the camera heading in radians.
+         * @memberof CameraController.prototype
          * @type {Number}
          */
         heading : {
@@ -852,9 +856,8 @@ define([
         },
 
         /**
-         * The the camera tilt in radians
-         * @memberof CameraController
-         *
+         * Gets or sets the camera tilt in radians
+         * @memberof CameraController.prototype
          * @type {Number}
          */
         tilt : {