+++ Source/Scene/Primitive.js
@@ -110,7 +110,7 @@ define([
      *     material : Cesium.Material.fromType('Checkerboard')
      *   })
      * });
-     * scene.getPrimitives().add(primitive);
+     * scene.primitives.add(primitive);
      *
      * // 2. Draw different instances each with a unique color
      * var extentInstance = new Cesium.GeometryInstance({
@@ -143,7 +143,7 @@ define([
      *   geometryInstances : [extentInstance, ellipsoidInstance],
      *   appearance : new Cesium.PerInstanceColorAppearance()
      * });
-     * scene.getPrimitives().add(primitive);
+     * scene.primitives.add(primitive);
      *
      * // 3. Create the geometry on the main thread.
      * var primitive = new Cesium.Primitive({
@@ -161,7 +161,7 @@ define([
      *   }),
      *   appearance : new Cesium.PerInstanceColorAppearance()
      * });
-     * scene.getPrimitives().add(primitive);
+     * scene.primitives.add(primitive);
      *
      * @see GeometryInstance
      * @see Appearance
@@ -721,7 +721,7 @@ define([
                 this.geometryInstances = undefined;
             }
 
-            this._geomtries = undefined;
+            this._geometries = undefined;
             this._createdGeometries = undefined;
             this._state = PrimitiveState.COMPLETE;
         }