+++ Specs/Scene/CameraControllerSpec.js
@@ -416,7 +416,7 @@ defineSuite([
         expect(camera.frustum.bottom).toEqual(-0.75, CesiumMath.EPSILON10);
     });
 
-    it('clamps zoom in 2D', function() {
+    it('clamps zoom out in 2D', function() {
         var frustum = new OrthographicFrustum();
         frustum.near = 1.0;
         frustum.far = 2.0;
@@ -440,6 +440,29 @@ defineSuite([
         expect(frustum.left).toBeGreaterThan(-dx);
         expect(frustum.top).toEqual(frustum.right * ratio);
         expect(frustum.bottom).toEqual(-frustum.top);
+    });
+
+    it('clamps zoom in in 2D', function() {
+        var frustum = new OrthographicFrustum();
+        frustum.near = 1.0;
+        frustum.far = 2.0;
+        frustum.left = -2.0;
+        frustum.right = 2.0;
+        frustum.top = 1.0;
+        frustum.bottom = -1.0;
+        camera.frustum = frustum;
+
+        var ellipsoid = Ellipsoid.WGS84;
+        var projection = new GeographicProjection(ellipsoid);
+        controller.update(SceneMode.SCENE2D, { projection : projection });
+
+        var max = projection.project(new Cartographic(Math.PI, CesiumMath.toRadians(85.05112878)));
+        var factor = 1000.0;
+        var dx = max.x * factor;
+
+        controller.zoomIn(dx * 2.0);
+        expect(frustum.right).toEqual(1.0);
+        expect(frustum.left).toEqual(-1.0);
     });
 
     it('zooms in 3D', function() {
@@ -793,7 +816,7 @@ defineSuite([
 
     it('pick ellipsoid', function() {
         var ellipsoid = Ellipsoid.WGS84;
-        var maxRadii = ellipsoid.getMaximumRadius();
+        var maxRadii = ellipsoid.maximumRadius;
 
         camera.position = Cartesian3.multiplyByScalar(Cartesian3.UNIT_X, 2.0 * maxRadii);
         camera.direction = Cartesian3.normalize(Cartesian3.negate(camera.position));
@@ -819,7 +842,7 @@ defineSuite([
     it('pick map in 2D', function() {
         var ellipsoid = Ellipsoid.WGS84;
         var projection = new GeographicProjection(ellipsoid);
-        var maxRadii = ellipsoid.getMaximumRadius();
+        var maxRadii = ellipsoid.maximumRadius;
 
         camera.position = new Cartesian3(0.0, 0.0, 2.0 * maxRadii);
         camera.direction = Cartesian3.normalize(Cartesian3.negate(camera.position));
@@ -849,7 +872,7 @@ defineSuite([
     it('pick rotated map in 2D', function() {
         var ellipsoid = Ellipsoid.WGS84;
         var projection = new GeographicProjection(ellipsoid);
-        var maxRadii = ellipsoid.getMaximumRadius();
+        var maxRadii = ellipsoid.maximumRadius;
 
         camera.position = new Cartesian3(0.0, 0.0, 2.0 * maxRadii);
         camera.direction = Cartesian3.normalize(Cartesian3.negate(camera.position));
@@ -885,7 +908,7 @@ defineSuite([
     it('pick map in columbus view', function() {
         var ellipsoid = Ellipsoid.WGS84;
         var projection = new GeographicProjection(ellipsoid);
-        var maxRadii = ellipsoid.getMaximumRadius();
+        var maxRadii = ellipsoid.maximumRadius;
 
         camera.position = Cartesian3.multiplyByScalar(Cartesian3.normalize(new Cartesian3(0.0, -1.0, 1.0)), 5.0 * maxRadii);
         camera.direction = Cartesian3.normalize(Cartesian3.subtract(Cartesian3.ZERO, camera.position));
@@ -925,7 +948,7 @@ defineSuite([
     it('set position cartographic in 2D', function() {
         var ellipsoid = Ellipsoid.WGS84;
         var projection = new GeographicProjection(ellipsoid);
-        var maxRadii = ellipsoid.getMaximumRadius();
+        var maxRadii = ellipsoid.maximumRadius;
 
         controller._mode = SceneMode.SCENE2D;
         controller._projection = projection;
@@ -1152,7 +1175,7 @@ defineSuite([
     it('gets magnitude in 2D', function() {
         var ellipsoid = Ellipsoid.WGS84;
         var projection = new GeographicProjection(ellipsoid);
-        var maxRadii = ellipsoid.getMaximumRadius();
+        var maxRadii = ellipsoid.maximumRadius;
 
         controller._mode = SceneMode.SCENE2D;
         controller._projection = projection;
@@ -1274,7 +1297,7 @@ defineSuite([
     });
 
     it('animates position to visible map in Columbus view', function() {
-        var maxRadii = Ellipsoid.WGS84.getMaximumRadius();
+        var maxRadii = Ellipsoid.WGS84.maximumRadius;
         var frustum = new PerspectiveFrustum();
         frustum.fovy = CesiumMath.toRadians(60.0);
         frustum.aspectRatio = context.getDrawingBufferWidth() / context.getDrawingBufferHeight();
@@ -1324,7 +1347,7 @@ defineSuite([
     });
 
     it('animates position to visible map in Columbus view with web mercator projection', function() {
-        var maxRadii = Ellipsoid.WGS84.getMaximumRadius();
+        var maxRadii = Ellipsoid.WGS84.maximumRadius;
         var frustum = new PerspectiveFrustum();
         frustum.fovy = CesiumMath.toRadians(60.0);
         frustum.aspectRatio = context.getDrawingBufferWidth() / context.getDrawingBufferHeight();