+++ Specs/Widgets/CesiumWidget/CesiumWidgetSpec.js
@@ -99,7 +99,7 @@ defineSuite([
             imageryProvider : new TileCoordinatesImageryProvider()
         };
         widget = new CesiumWidget(container, options);
-        var imageryLayers = widget.centralBody.getImageryLayers();
+        var imageryLayers = widget.centralBody.imageryLayerCollection;
         expect(imageryLayers.getLength()).toEqual(1);
         expect(imageryLayers.get(0).getImageryProvider()).toBe(options.imageryProvider);
     });
@@ -108,7 +108,7 @@ defineSuite([
         widget = new CesiumWidget(container, {
             imageryProvider : false
         });
-        var imageryLayers = widget.centralBody.getImageryLayers();
+        var imageryLayers = widget.centralBody.imageryLayerCollection;
         expect(imageryLayers.getLength()).toEqual(0);
     });
 
@@ -155,7 +155,7 @@ defineSuite([
             contextOptions : contextOptions
         });
 
-        var context = widget.scene.getContext();
+        var context = widget.scene.context;
         var contextAttributes = context._gl.getContextAttributes();
 
         expect(context.options.allowTextureFilterAnisotropic).toEqual(false);